fix: Measure items properly in dual page viewmode
